Merged PR 170: Refactor CatalogSnapshot JSON handling for better performance
This change refactors how Machines and MachineGroups within a CatalogSnapshot are handled. Previously, too much reliance was placed on the MachineJsonResponse and MachineGroupJsonResponse JSON strings, pulled from OPDB and stored as-is in the database, to generate CatalogSnapshotDigest summaries as they were needed. This required deserializing the massive MachineJsonResponse and MachineGroupJsonResponse every time, which is an extremely expensive operation that caused API response times numbering in the tens of seconds.

Now, the deserialization only happens once when the snapshot is imported, and the results (along with digest information such as machine counts, machine group counts, manufacturer counts and the newest machine in a snapshot) are stored in the database as strongly-typed JSON columns called Machines and MachineGroups. MachineJsonResponse and MachineGroupJsonResponse are still stored, but they are no longer used for anything other than as a backup of the original Json responses. All application queries now use the Machines and MachineGroups columns, as well as the new Digest columns. Now, API response times are measured in milliseconds. Win.

Related work items: #352