Add new migration scripts and refactor Pinball data structure
This commit introduces new migration scripts for the database and refactors the existing Pinball data structure. The changes include the introduction of new tables, modifications to existing tables, as well as the dropping of obsolete tables. It also includes changes to accommodate new database schema for enhanced Pinball data management.