Add SqlCmd installation and SQL Azure deployment tasks to pipeline
Two new tasks have been added to the Azure pipeline: the installation of SqlCmd, and the deployment of an SQL Azure Dacpac. The SqlCmd tool is now installed via PowerShell, and the Deployment task is set up with appropriate inputs for a servicePrincipal authentication, server and database names, deployment type, a sql migration script to run, and IP detection method.